requirements.txt
nextcord
groq
requests
python-dotenv
aiosqlite